requirements.txt
python-json-logger==2.0.4
prometheus-client==0.17.0
botocore==1.21.32
googleapis-common-protos==1.53.0
opentelemetry-api==1.7.1
opentelemetry-exporter-otlp==1.7.1
opentelemetry-exporter-otlp-proto-grpc==1.7.1
opentelemetry-distro==0.26b1
opentelemetry-instrumentation==0.26b1
opentelemetry-instrumentation-boto==0.26b1
opentelemetry-instrumentation-botocore==0.26b1
opentelemetry-instrumentation-dbapi==0.26b1
opentelemetry-instrumentation-sqlite3==0.26b1
opentelemetry-instrumentation-wsgi==0.26b1
opentelemetry-proto==1.7.1
opentelemetry-sdk==1.7.1
opentelemetry-semantic-conventions==0.26b1
protobuf==3.17.3
